pyproject.toml
[tool.poetry]
name = "motorvibration"
version = "0.1.0"
description = ""
authors = ["Nagi0 <brunolefonseca13@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
polars = "^1.16.0"
python-dotenv = "^1.0.1"
tqdm = "^4.67.1"
scikit-learn = "^1.5.2"
scipy = "^1.14.1"
pyarrow = "^18.1.0"
pandas = "^2.2.3"
ipykernel = "^6.29.5"
matplotlib = "^3.9.3"
seaborn = "^0.13.2"
plotly = "^5.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
